pyproject.toml
[project]
name = "minecraft-world-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"colorama>=0.4.6",
	"python-dotenv",
]
